+++ Engine/Engine/Source/GameObjects/Circle.cpp
@@ -10,8 +10,8 @@ Circle::Circle(Vector2 spawn_pos, XEngine& ref, float32 _radiusX, float32 _radiu
 }
 
 void Circle::OnRender(Renderer &renderer) {
-	SetTransform(renderer, radiusX * 2.f, radiusY * 2.f);
-	renderer.RenderCircle(transform->position.x, transform->position.y, radiusX, radiusY, color, transform->scale, fill, strokeWith);
+	SetTransform(renderer,(int) (radiusX * 2.f), (int) (radiusY * 2.f));
+	renderer.RenderCircle(transform->position.x, transform->position.y, (float) radiusX, (float) radiusY, color, transform->scale, fill, strokeWith);
 	GameObject::OnRender(renderer);
 }
